--- examples/backgroundExample/App.js
@@ -1,140 +1,0 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- *
- * @format
- * @flow
- */
-
-import React from 'react';
-import {
-    SafeAreaView,
-    StyleSheet,
-    ScrollView,
-    View,
-    Text,
-    StatusBar,
-    TouchableOpacity,
-    Platform,
-    Linking,
-} from 'react-native';
-import { Header, Colors } from 'react-native/Libraries/NewAppScreen';
-
-import BackgroundJob from 'react-native-background-actions';
-
-const sleep = (time) => new Promise((resolve) => setTimeout(() => resolve(), time));
-
-BackgroundJob.on('expiration', () => {
-    console.log('iOS: I am being closed!');
-});
-
-const taskRandom = async (taskData) => {
-    if (Platform.OS === 'ios') {
-        console.warn(
-            'This task will not keep your app alive in the background by itself, use other library like react-native-track-player that use audio,',
-            'geolocalization, etc. to keep your app alive in the background while you excute the JS from this library.'
-        );
-    }
-    await new Promise(async (resolve) => {
-        // For loop with a delay
-        const { delay } = taskData;
-        console.log(BackgroundJob.isRunning(), delay)
-        for (let i = 0; BackgroundJob.isRunning(); i++) {
-            console.log('Runned -> ', i);
-            await BackgroundJob.updateNotification({ taskDesc: 'Runned -> ' + i });
-            await sleep(delay);
-        }
-    });
-};
-
-const options = {
-    taskName: 'Example',
-    taskTitle: 'ExampleTask title',
-    taskDesc: 'ExampleTask desc',
-    taskIcon: {
-        name: 'ic_launcher',
-        type: 'mipmap',
-    },
-    color: '#ff00ff',
-    linkingURI: 'exampleScheme://chat/jane',
-    parameters: {
-        delay: 1000,
-    },
-};
-
-function handleOpenURL(evt) {
-    console.log(evt.url);
-    // do something with the url
-}
-
-Linking.addEventListener('url', handleOpenURL);
-
-class App extends React.Component {
-    playing = BackgroundJob.isRunning();
-
-    /**
-     * Toggles the background task
-     */
-    toggleBackground = async () => {
-        this.playing = !this.playing;
-        if (this.playing) {
-            try {
-                console.log('Trying to start background service');
-                await BackgroundJob.start(taskRandom, options);
-                console.log('Successful start!');
-            } catch (e) {
-                console.log('Error', e);
-            }
-        } else {
-            console.log('Stop background service');
-            await BackgroundJob.stop();
-        }
-    };
-    render() {
-        return (
-            <>
-                <StatusBar barStyle="dark-content" />
-                <SafeAreaView>
-                    <ScrollView
-                        contentInsetAdjustmentBehavior="automatic"
-                        style={styles.scrollView}>
-                        <Header />
-                        {global.HermesInternal == null ? null : (
-                            <View style={styles.engine}>
-                                <Text style={styles.footer}>Engine: Hermes</Text>
-                            </View>
-                        )}
-                        <View style={styles.body}>
-                            <TouchableOpacity
-                                style={{ height: 100, width: 100, backgroundColor: 'red' }}
-                                onPress={this.toggleBackground}></TouchableOpacity>
-                        </View>
-                    </ScrollView>
-                </SafeAreaView>
-            </>
-        );
-    }
-}
-
-const styles = StyleSheet.create({
-    scrollView: {
-        backgroundColor: Colors.lighter,
-    },
-    engine: {
-        position: 'absolute',
-        right: 0,
-    },
-    body: {
-        backgroundColor: Colors.white,
-    },
-    footer: {
-        color: Colors.dark,
-        fontSize: 12,
-        fontWeight: '600',
-        padding: 4,
-        paddingRight: 12,
-        textAlign: 'right',
-    },
-});
-
-export default App;
